com: more networking fixes

+++ com/src/com.h
@@ -5,7 +5,7 @@
 
 typedef struct {
   bool networked;
-  uint16_t generation;
+  uint32_t generation;
 } naos_com_status_t;
 
 typedef struct {
